# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "老年人用药管理系统",
    "description": "老人用药管理智能助手后端 API",
    "version": "2.33.0"
  },
  "paths": {
...
    "/api/v1/ai/chat": {
      "post": {
        "tags": [
          "AI健康助手"
        ],
        "summary": "Chat",
        "description": "向AI健康助手提问（需要认证，使用当前用户各自的 AI 配置）",
        "operationId": "chat_api_v1_ai_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas__ai__AIQuestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAnswer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ai/chat/public": {
      "post": {
        "tags": [
          "AI健康助手"
        ],
        "summary": "Chat Public",
        "description": "向AI健康助手提问（无需认证，供老人端使用，基于 IP 限流）",
        "operationId": "chat_public_api_v1_ai_chat_public_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas__ai__AIQuestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAnswer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/user/ai-config": {
      "get": {
        "tags": [
          "AI健康助手配置"
        ],
        "summary": "Get Ai Config",
        "description": "读取 AI 配置（api_key 不回传明文，仅返回 has_api_key 标识）",
        "operationId": "get_ai_config_api_v1_ai_user_ai_config_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "目标设备ID（家属为被照护老人配置）",
              "title": "Device Id"
            },
            "description": "目标设备ID（家属为被照护老人配置）"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "目标用户ID",
              "title": "User Id"
            },
            "description": "目标用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAIConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AI健康助手配置"
        ],
        "summary": "Upsert Ai Config",
        "description": "新增或更新 AI 配置（api_key 加密存储）",
        "operationId": "upsert_ai_config_api_v1_ai_user_ai_config_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "目标设备ID（家属为被照护老人配置）",
              "title": "Device Id"
            },
            "description": "目标设备ID（家属为被照护老人配置）"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "目标用户ID",
              "title": "User Id"
            },
            "description": "目标用户ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAIConfigIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAIConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/providers": {
      "get": {
        "tags": [
          "AI健康助手配置"
        ],
        "summary": "List Providers",
        "description": "公开：返回前端下拉所需的厂商预设列表",
        "operationId": "list_providers_api_v1_ai_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AIProviderPreset"
                  },
                  "type": "array",
                  "title": "Response List Providers Api V1 Ai Providers Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIAnswer": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "AIAnswer",
        "description": "AI 回答响应"
      },
      "AIProviderPreset": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "default_model": {
            "type": "string",
            "title": "Default Model",
            "default": ""
          },
          "requires_base_url": {
            "type": "boolean",
            "title": "Requires Base Url",
            "default": false
          },
          "requires_model": {
            "type": "boolean",
            "title": "Requires Model",
            "default": true
          },
          "doc_url": {
            "type": "string",
            "title": "Doc Url",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "AIProviderPreset",
        "description": "前端厂商预设（用于下拉选择）"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserAIConfigIn": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "zhipuai"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "title": "UserAIConfigIn",
        "description": "每用户 AI 配置（写入）"
      },
      "UserAIConfigOut": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "has_api_key": {
            "type": "boolean",
            "title": "Has Api Key",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "enabled"
        ],
        "title": "UserAIConfigOut",
        "description": "每用户 AI 配置（返回，api_key 不回传明文，仅告知是否已配置）"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__schemas__ai__AIQuestion": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AIQuestion",
        "description": "AI 健康助手提问请求"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}